chore(jans-cedarling): refactor move parsing `CedarSchema` closer to the structure
