+++ jans-cedarling/cedarling/src/init/mod.rs
@@ -10,6 +10,5 @@
 //! - load Cedar Policies
 //! - get keys for JWT validation
 
-pub(crate) mod cedar_schema;
 pub(crate) mod policy;
 pub(crate) mod policy_store;
