+++ jans-cedarling/test_files/policy-store_ok/cedar.schema
@@ -1,0 +1,13 @@
+namespace Jans {
+type Url = {"host": String, "path": String, "protocol": String};
+entity Issue = {"org_id": String};
+entity TrustedIssuer = {"issuer_entity_id": Url};
+entity Workload = {"client_id": String, "iss": TrustedIssuer, "name": String, "org_id": String};
+entity Access_token = {"aud": String, "exp": Long, "iat": Long, "iss": TrustedIssuer, "jti": String};
+action "Update" appliesTo {
+  principal: [Workload],
+  resource: [Issue],
+  context: {}
+};
+}
+
